+++ geo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `geo-utils` CHANGELOG
 
+## 1.2.0
+
+- New functions:
+  - `isValidLat(lat)` and `isValidLon(lon)`
+  - `isClose(pointA, pointB, threshold)`, to check if two points are close enough
+
 ## 1.1.1
 
 - Fix JSDoc using wrong lib name (NumberUtils).